new branch to move parseBoard into own file
