+++ app/main.py
@@ -1,6 +1,7 @@
 import json
 import os
 import bottle
+from parseBoard import setBoardValues
 
 from api import ping_response, start_response, move_response, end_response
 from pathfinding.core.grid import Grid
@@ -137,45 +138,7 @@ def boardToArray(dataDump):
     return board
 
 
-def setBoardValues(jData):
-    """
-    Converts converted JSON data from the battlesnake engine, to an a_star friendly gameboard.
-    Args:
-        jData (list): Converted JSON data
 
-    Returns:
-        A_Star friendly version of the gameboard.
-
-    """
-    board = setEdge(jData)
-
-    # pertaining to our own body
-    me = jData["you"]["id"]
-
-    for z in jData["you"]["body"]:
-        if (z == jData["you"]["body"][0]):
-            x = z['x']
-            y = z['y']
-            board[y][x] = ourHead
-        else:
-            x = z['x']
-            y = z['y']
-            board[y][x] = SNAKE
-
-    # other snakes
-    for z in jData["board"]["snakes"]:
-        name = z["id"]
-        for a in z["body"]:
-            if (name != me):
-                if (a == z["body"][0]):
-                    x = a['x']
-                    y = a['y']
-                    board[y][x] = SNAKE
-                else:
-                    x = a['x']
-                    y = a['y']
-                    board[y][x] = SNAKE
-    return board
 
 
 def setEdge(dataDump):
